pyproject.toml
[project]
name = "text-to-code-rag"
version = "0.1.0"
description = "Semantic search system for Python code fragments using bi-encoder architecture"
authors = [
    {name = "Arseny Pivovarov", email = "pivovarovsenya@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch==2.2.2",
    "torchvision>=0.17.0,<0.18.0",
    "pytorch-lightning>=2.1.0,<2.3.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "mlflow>=2.8.0",
    "dvc>=3.30.0",
    "fire>=0.5.0",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "tokenizers>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.5.0",
    "ruff>=0.1.6",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
text-to-code-rag = "text_to_code_rag.commands:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["text_to_code_rag"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=text_to_code_rag --cov-report=term-missing"
